src/whisper/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

git+https://github.com/openai/whisper.git
ruamel.yaml
srt
rich
librosa
